+++ src/profile/mod.rs
@@ -1,7 +1,9 @@
 mod convert;
 mod frame;
+mod perfetto;
 mod sample;
 
 pub use convert::*;
 pub use frame::*;
+pub use perfetto::*;
 pub use sample::*;